Bench.Cardano.Ledger.EpochBoundary

Description

Benchmarks for things which happen on an epoch boundary.

Documentation

type TestEra = MaryEra TestCrypto

payCred ∷ PaymentCredential TestCrypto

txIns ∷ [TxIn TestCrypto]

Infinite list of transaction inputs

txOutUnstaked ∷ TxOut TestEra

Unstaked address

txOutsFromCreds ∷ [StakeCredential TestCrypto] → [TxOut TestEra]

Generate TxOuts for each stake credential.

txOutsFromPtrs ∷ [Ptr] → [TxOut TestEra]

stakeCreds ∷ Word64 → [StakeCredential TestCrypto]

Generate n stake credentials

stakePtrs ∷ [StakeCredential c] → Map Ptr (StakeCredential c)

Generate pointers to a list of stake credentials

utxo

Arguments

∷ Word64
Number of unstaked outputs

→ Word64
Number of distinct base stake credentials

→ Map Ptr c

→ Int

→ UTxO TestEra

Create a UTxO set containing noUnstaked unstaked TxOuts, noBase base stake credentials and any pointer credentials passed in ptrMap. For each credential, there will be dupFactor entries in the UTxO.

data AggTestSetup

Constructors

AggTestSetup

Fields

atsPtrMap ∷ !(Map Ptr (StakeCredential TestCrypto))

atsUTxO ∷ !(UTxO TestEra)

sizedAggTestSetup

Arguments

∷ Word64
Number of unstaked addresses

→ Word64
Number of distinct base stake credentials

→ Word64
Number of distinct pointer stake credentials

→ Int
Duplication factor

→ AggTestSetup

Construct the relevant UTxO and pointer map to test aggregateUtxoCoinByCredential.

There are three parameters to this: - The number of distinct StakeRefBase outputs present in the UTxO. - The number of distinct StakeRefPtr outputs present in the UTxO, and the corresponding size of the pointer map. - The "duplication factor". This represents the number of TxIns pointing to outputs with a given stake address. Thus with a duplication factor of 1, there will be only one "coin" per credential, and little aggregation. With a factor of 10, there will be 10 coins that must be aggregated.

Note that this corresponds to two situations in real usage: multiple transactions paying to the same address, and multiple addresses sharing the same stake credential. These scenarios are identical for benchmarking purposes, and hence we roll them into one.

aggregateUtxoBench ∷ Benchmark