Cardano.Chain.Byron.API.Common

Extract info from genesis config

allowedDelegators ∷ Config → Set KeyHash

Extract info from chain state

getDelegationMap ∷ ChainValidationState → Map

getProtocolParams ∷ ChainValidationState → ProtocolParameters

getMaxBlockSize ∷ ChainValidationState → Word32

Annotations

reAnnotateBlock ∷ EpochSlots → ABlock () → ABlock ByteString

reAnnotateBoundary ∷ ProtocolMagicId → ABoundaryBlock () → ABoundaryBlock ByteString

reAnnotateMagic ∷ ProtocolMagic → AProtocolMagic ByteString

reAnnotateMagicId ∷ ProtocolMagicId → Annotated ProtocolMagicId ByteString

reAnnotateUsing ∷ ∀ f a. Functor f ⇒ (f a → Encoding) → (∀ s. Decoder s (f ByteSpan)) → f a → f ByteString

Generalization of reAnnotate

Headers

abobMatchesBody ∷ ABlockOrBoundaryHdr ByteString → ABlockOrBoundary ByteString → Bool

Check if a block matches its header

For EBBs, we're currently being more permissive here and not performing any header-body validation but only checking whether an EBB header and EBB block were provided. This seems to be fine as it won't cause any loss of consensus with the old `cardano-sl` nodes.