Cardano.Chain.Common.Attributes

Description

Helper data type for block, tx attributes

Map with integer 1-byte keys, arbitrary-type polymorph values. Needed primarily for partial serialization. Values are either parsed and put to some constructor or left as unparsed.

Documentation

newtype UnparsedFields

Representation of unparsed fields in Attributes. Newtype wrapper is used for clear backward compatibility between previous representation (which was just a single ByteString) during transition from Store to CBOR.

Constructors

UnparsedFields (Map Word8 ByteString)

data Attributes h

Convenient wrapper for the datatype to represent it (in binary format) as k-v map

Constructors

Attributes

Fields

attrData ∷ h

Data, containing known keys (deserialized)

attrRemain ∷ UnparsedFields

Remaining, unparsed fields

attributesAreKnown ∷ Attributes a → Bool

Check whether all data from Attributes is known, i. e. was successfully parsed into some structured data

unknownAttributesLength ∷ Attributes a → Int

toCBORAttributes ∷ ∀ t. [(Word8, t → ByteString)] → Attributes t → Encoding

fromCBORAttributes ∷ ∀ t s. t → (Word8 → ByteString → t → Decoder s (Maybe t)) → Decoder s (Attributes t)

mkAttributes ∷ h → Attributes h

dropAttributes ∷ Dropper s

dropEmptyAttributes ∷ Dropper s

Drop `Attributes ()` making sure that the UnparsedFields are empty