Cardano.Chain.Genesis.KeyHashes

Documentation

newtype GenesisKeyHashes

The set of genesis keys, who are able to produce blocks and submit votes and proposals in the Byron era

Constructors

GenesisKeyHashes

Fields

unGenesisKeyHashes ∷ Set KeyHash