Cardano.Chain.ProtocolConstants

Description

Constants derived from security parameter

TODO: Find a better home for these

Documentation

kSlotSecurityParam ∷ BlockCount → SlotCount

Security parameter expressed in number of slots. It uses chain quality property. It's basically blkSecurityParam / chainQualityThreshold.

kUpdateStabilityParam ∷ BlockCount → SlotCount

Update stability parameter expressed in number of slots. This is the time between an protocol version update receiving its final endorsement and being accepted, and is set to double the security param.

This extra safety margin is required because an update in the protocol version may trigger a hard fork, which can change "era"-level parameters such as slot length and the number of slots per epoch. As such, the consensus layer wishes to always have a margin between such an update being _certain to happen_ and it actually happening.

For full details, you can see https://github.com/input-output-hk/cardano-ledger-specs/issues/1288

kChainQualityThreshold ∷ Fractional f ⇒ BlockCount → f

Minimal chain quality (number of blocks divided by number of slots) necessary for security of the system.

kEpochSlots ∷ BlockCount → EpochSlots

Number of slots inside one epoch