Cardano.Chain.Update.Validation.Endorsement

Documentation

data Environment

Constructors

Environment

Fields

k ∷ !BlockCount

Chain stability parameter.

currentSlot ∷ !SlotNumber

adoptionThreshold ∷ !Int

delegationMap ∷ !Map

confirmedProposals ∷ !(Map UpId SlotNumber)

registeredProtocolUpdateProposals ∷ !ProtocolUpdateProposals

data State

Constructors

State

Fields

candidateProtocolVersions ∷ ![CandidateProtocolUpdate]

registeredEndorsements ∷ !(Set Endorsement)

data Endorsement

Constructors

Endorsement

Fields

endorsementProtocolVersion ∷ !ProtocolVersion

endorsementKeyHash ∷ !KeyHash

data CandidateProtocolUpdate

Constructors

CandidateProtocolUpdate

Fields

cpuSlot ∷ !SlotNumber

Slot at which this protocol version and parameters gathered enough endorsements and became a candidate. This is used to check which versions became candidates 2k slots before the end of an epoch (and only those can be adopted at that epoch). Versions that became candidates later than 2k slots before the end of an epoch can be adopted in following epochs.

cpuProtocolVersion ∷ !ProtocolVersion

cpuProtocolParameters ∷ !ProtocolParameters

register ∷ MonadError Error m ⇒ Environment → State → Endorsement → m State

Register an endorsement.

This corresponds to the UPEND rule.

data Error

Constructors

MultipleProposalsForProtocolVersion ProtocolVersion
Multiple proposals were found, which propose an update to the same protocol version.