Cardano.Chain.Update.Validation.Interface.ProtocolVersionBump

Documentation

data Environment

Constructors

Environment

Fields

k ∷ !BlockCount

epochFirstSlot ∷ !SlotNumber

candidateProtocolVersions ∷ ![CandidateProtocolUpdate]

data State

Constructors

State

Fields

nextProtocolVersion ∷ !ProtocolVersion

nextProtocolParameters ∷ !ProtocolParameters

tryBumpVersion ∷ Environment → State → State

Change the protocol version when an epoch change is detected, and there is a candidate protocol update that was confirmed at least 4 * k slots before the start of the new epoch, where k is the chain security parameter.

For a full history of why this is required, see https://github.com/input-output-hk/cardano-ledger-specs/issues/1288

This corresponds to the PVBUMP rules in the Byron ledger specification.