Constrained.Test

Documentation

testAll ∷ IO ()

tests ∷ Bool → Spec

negativeTests ∷ Spec

numberyTests ∷ Spec

sizeTests ∷ Spec

data NumberyType where

Constructors

N ∷ (Typeable a, Numbery a) ⇒ Proxy a → NumberyType

testNumberyListSpec ∷ String → (∀ a. Numbery a ⇒ Specification BaseFn [a]) → Spec

testNumberyListSpecNoShrink ∷ String → (∀ a. Numbery a ⇒ Specification BaseFn [a]) → Spec

testNumberyListSpec' ∷ Bool → String → (∀ a. Numbery a ⇒ Specification BaseFn [a]) → Spec

testSpec ∷ HasSpec fn a ⇒ String → Specification fn a → Spec

testSpecNoShrink ∷ HasSpec fn a ⇒ String → Specification fn a → Spec

testSpec' ∷ HasSpec fn a ⇒ Bool → String → Specification fn a → Spec

noInfinity ∷ Gen (NumSpec fn Integer)

When we multiply intervals, we get a bounding box, around the possible values. When the intervals have infinities, the bounding box can be very loose. In fact the order in which we multiply intervals with infinities can affect how loose the bounding box is. So ((NegInf, n) * (a, b)) * (c,d) AND (NegInf, n) * ((a, b) * (c,d)) may have different bounding boxes To test the associative laws we must have no infinities, and then the associative law will hold.

plusNegate ∷ NumSpec fn Integer → NumSpec fn Integer → Property

commutesNumSpec ∷ NumSpec fn Integer → NumSpec fn Integer → Property

assocNumSpec ∷ NumSpec fn Integer → NumSpec fn Integer → NumSpec fn Integer → Property

commuteTimes ∷ NumSpec fn Integer → NumSpec fn Integer → Property

assocNumSpecTimes ∷ Gen Property

negNegate ∷ NumSpec fn Integer → Property

scaleNumSpec ∷ NumSpec fn Integer → Property

scaleOne ∷ NumSpec fn Integer → Property

numNumSpecTree ∷ Spec

hasSizeList ∷ Specification BaseFn [Int]

hasSizeSet ∷ Specification BaseFn (Set Int)

hasSizeMap ∷ Specification BaseFn (Map Int Int)