GetDataFileName

Documentation

withTestFileProperty ∷ FilePath → (FilePath → Property) → Property

This contraption is necessary because test modules in the folder are used in two different packages, but the file paths for test files are relative to the package itself, which has to accounted for.

(<:<) ∷ (FilePath → Property) → FilePath → Property