Main

Documentation

genVl ∷ Gen Coin

data BenchCrypto

type BenchEra = ShelleyEra BenchCrypto

type FixedValType = Coin

eqf ∷ String → (Map Int Int → Map Int Int → Bool) → Int → Benchmark

mainEq ∷ IO ()

includes_init_SpendOneUTxO ∷ IO ()

profileUTxO ∷ IO ()

touchCertState ∷ CertState c → Int

touchUTxOState ∷ UTxOState cryto → Int

profileCreateRegKeys ∷ IO ()

profileNkeysMPools ∷ IO ()

profileCreateRegPools ∷ Word64 → IO ()

profileEpochBoundary ∷ Benchmark

epochAt ∷ Int → Benchmark

action2m ∷ EraTxOut era ⇒ (DState era, PState era, UTxO era) → SnapShot (EraCrypto era)

action2im ∷ ∀ era. EraTxOut era ⇒ (DState era, PState era, UTxO era) → SnapShot (EraCrypto era)

validGroup ∷ Benchmark

Benchmarks for the various validation transitions exposed by the API

profileValid ∷ IO ()

domainRangeRestrict ∷ IO ()

drrAt ∷ Int → Benchmark

alg1 ∷ (Map Int Int, Map Int Char, Map Char Char) → Map Int Char

alg2 ∷ (Map Int Int, Map Int Char, Map Char Char) → Map Int Char

varyState ∷ NFData state ⇒ String → Word64 → [Word64] → (Word64 → Word64 → state) → (Word64 → Word64 → state → ()) → Benchmark

varyInput ∷ NFData state ⇒ String → (Word64, Word64) → [(Word64, Word64)] → (Word64 → Word64 → state) → (Word64 → Word64 → state → ()) → Benchmark

varyDelegState ∷ NFData state ⇒ String → Word64 → [Word64] → (Word64 → Word64 → state) → (Word64 → Word64 → state → ()) → Benchmark

main ∷ IO ()