Test.Byron.Spec.Ledger.Delegation.Properties

Description

Properties of the delegation traces induced by the transition systems associated with this aspect of the ledger.

Documentation

dcertsAreTriggered ∷ Property

rejectDupSchedDelegs ∷ Property

Reject delegation certificates where a genesis key tries to delegate in the same slot.

This property tries to generate a trace where the last state contains a non-empty sequence of scheduled delegations. If such trace cannot be generated, then the test will fail when the heap limit is reached, or hedgehog gives up.

tracesAreClassified ∷ Property

Classify the traces.

dblockTracesAreClassified ∷ Property

relevantCasesAreCovered ∷ Property

onlyValidSignalsAreGenerated ∷ Property