Test.Byron.Spec.Ledger.UTxO.Properties

Documentation

moneyIsConstant ∷ Property

Check that the money is constant in the system.

noDoubleSpending ∷ Property

Check that there is no double spending

utxoDiff ∷ Property

Check that UTxO is outputs minus inputs

utxoAndTxoutsMustBeDisjoint ∷ Property

relevantCasesAreCovered ∷ Property

avgInputsOutputs ∷ [TxBody] → (Double, Double)

Returns the average number of inputs and outputs for a list of transactions.

tracesAreClassified ∷ Property