Test.Cardano.Ledger.Shelley.Examples

Documentation

data CHAINExample h era

Constructors

CHAINExample

Fields

startState ∷ ChainState era

State to start testing with

newBlock ∷ Block h era

Block to run chain state transition system on

intendedResult ∷ Either (NonEmpty (PredicateFailure (CHAIN era))) (ChainState era)

type of fatal error, if failure expected and final chain state if success expected

testCHAINExample ∷ HasCallStack ⇒ CHAINExample (BHeader C_Crypto) C → Assertion

Runs example, applies chain state transition system rule (STS), and checks that trace ends with expected state or expected error.