Test.Cardano.Ledger.Shelley.Examples.Combinators

Description

A collection of combinators for manipulating Chain State. The idea is to provide a clear way of describing the changes to the chain state when a block is processed.

Documentation

evolveNonceFrozen ∷ ∀ era. Nonce → ChainState era → ChainState era

Evolve Nonces - Frozen

Evolve the appropriate nonces under the assumption that the candidate nonce is now frozen.

evolveNonceUnfrozen ∷ ∀ era. Nonce → ChainState era → ChainState era

Evolve Nonces - Unfrozen

Evolve the appropriate nonces under the assumption that the candidate nonce is not frozen. Note: do not use this function when crossing the epoch boundary, instead use newEpoch.

newLab ∷ ∀ era. Era era ⇒ Block (BHeader (EraCrypto era)) era → ChainState era → ChainState era

New LastAppliedBlock (*NOT* on epoch boundaries)

Update the chain state with the details of LastAppliedBlock that occur when a new block is processed. Note: do not use this function when crossing the epoch boundary, instead use newEpoch.

feesAndKeyRefund ∷ ∀ era. Coin → Credential 'Staking (EraCrypto era) → ChainState era → ChainState era

feesAndDeposits ∷ ∀ era. EraPParams era ⇒ PParams era → Coin → [Credential 'Staking (EraCrypto era)] → [PoolParams (EraCrypto era)] → ChainState era → ChainState era

Update Fees and Deposits

Update the fee pot and deposit pot with the new fees and deposits adjust the deposit tables in the UTxOState and the CertState. Notes 1) do not give this function duplicates in the stakes or pools inputs. 2) do not use this function when crossing the epoch boundary, instead use newEpoch.

newUTxO ∷ ∀ era. (EraTx era, EraGov era) ⇒ TxBody era → ChainState era → ChainState era

Update the UTxO

Update the UTxO for given transaction body.

newStakeCred ∷ ∀ era. Credential 'Staking (EraCrypto era) → Ptr → ChainState era → ChainState era

New Stake Credential

Add a newly registered stake credential, initialize the rdRewards component of the RDPair. The rdDeposit component of the RDPair is set by feesAndDeposits

deregStakeCred ∷ ∀ era. Credential 'Staking (EraCrypto era) → ChainState era → ChainState era

De-Register Stake Credential

De-register a stake credential and all associated data. Be sure to run feesAndKeyRefund before you run this because this throws away the stored refund, which then can't be used to balance the utxosDeposited field in feesAndKeyRefund

delegation ∷ ∀ era. Credential 'Staking (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → ChainState era → ChainState era

New Delegation

Create a delegation from the given stake credential to the given stake pool.

newPool ∷ ∀ era. PoolParams (EraCrypto era) → ChainState era → ChainState era

New Stake Pool

Add a newly registered stake pool

reregPool ∷ ∀ era. PoolParams (EraCrypto era) → ChainState era → ChainState era

Re-Register Stake Pool

updatePoolParams ∷ ∀ era. PoolParams (EraCrypto era) → ChainState era → ChainState era

Re-Register Stake Pool

stageRetirement ∷ ∀ era. KeyHash 'StakePool (EraCrypto era) → EpochNo → ChainState era → ChainState era

Pool Retirement

Stage a stake pool for retirement.

reapPool ∷ ∀ era. EraGov era ⇒ PoolParams (EraCrypto era) → ChainState era → ChainState era

Reap Pool

Remove a stake pool.

mir ∷ ∀ era. Credential 'Staking (EraCrypto era) → MIRPot → Coin → ChainState era → ChainState era

MIR

Add a credential to the MIR mapping for the given pot (reserves or treasury)

applyMIR ∷ ∀ era. MIRPot → Map (Credential 'Staking (EraCrypto era)) Coin → ChainState era → ChainState era

Apply MIR

On the epoch boundary, reset the MIR mappings and augment the rewards.

rewardUpdate ∷ ∀ era. RewardUpdate (EraCrypto era) → ChainState era → ChainState era

Reward Update

Update the chain state with the given reward update

pulserUpdate ∷ ∀ era. PulsingRewUpdate (EraCrypto era) → ChainState era → ChainState era

Pulser

Update the chain state with the given reward update pulser

applyRewardUpdate ∷ ∀ era. EraGov era ⇒ RewardUpdate (EraCrypto era) → ChainState era → ChainState era

Apply a Reward Update

Apply the given reward update to the chain state

setPoolDistr ∷ ∀ era. PoolDistr (EraCrypto era) → ChainState era → ChainState era

Set Pool Distribution

Set the stake pool distribution to the given one.

setOCertCounter ∷ ∀ era. KeyHash 'BlockIssuer (EraCrypto era) → Word64 → ChainState era → ChainState era

Set Operation Certificate Counter

Set the operational certificates counter for a given stake pool.

newSnapshot ∷ ∀ era. SnapShot (EraCrypto era) → Coin → ChainState era → ChainState era

New Snapshot

Add a new snapshot and rotate the others

incrBlockCount ∷ ∀ era. KeyHash 'StakePool (EraCrypto era) → ChainState era → ChainState era

Increase Block Count

Record that the given stake pool (non-core node) produced a block.

newEpoch ∷ ∀ era. (ProtVerAtMost era 6, EraGov era) ⇒ Block (BHeader (EraCrypto era)) era → ChainState era → ChainState era

New Epoch

Update the new epoch number, set the nonces, set the last applied block, and reset blocks made. Note: This function subsumes the manipulations done by newLab, evolveNonceUnfrozen, and evolveNonceFrozen.

setCurrentProposals ∷ ∀ era. (GovState era ~ ShelleyGovState era, EraPParams era) ⇒ ProposedPPUpdates era → ChainState era → ChainState era

Set Current Proposals

Set the current protocol parameter proposals.

setFutureProposals ∷ ∀ era. GovState era ~ ShelleyGovState era ⇒ ProposedPPUpdates era → ChainState era → ChainState era

Set Future Proposals

Set the future protocol parameter proposals.

solidifyProposals ∷ ∀ era. EraGov era ⇒ ChainState era → ChainState era

setPParams ∷ ∀ era. EraGov era ⇒ PParams era → ChainState era → ChainState era

Set the Protocol Proposals

Set the protocol parameters.

setPrevPParams ∷ ∀ era. EraGov era ⇒ PParams era → ChainState era → ChainState era

Set the Previous Protocol Proposals

Set the previous protocol parameters.

setFutureGenDeleg ∷ ∀ era. (FutureGenDeleg (EraCrypto era), GenDelegPair (EraCrypto era)) → ChainState era → ChainState era

Set a future genesis delegation.

adoptFutureGenDeleg ∷ ∀ era. (FutureGenDeleg (EraCrypto era), GenDelegPair (EraCrypto era)) → ChainState era → ChainState era

Set a future genesis delegation.