Test.Cardano.Ledger.Shelley.Examples.EmptyBlock

Documentation

exEmptyBlock ∷ (EraBlockBody era, EraGov era, EraStake era, EraCertState era, AtMostEra "Mary" era, AtMostEra "Alonzo" era, Default (StashedAVVMAddresses era)) ⇒ CHAINExample era

Empty Block Example

This is the most minimal example of using the CHAIN STS transition. It applies an empty block to an initial shelley chain state.

The only things that change in the chain state are the evolving and candidate nonces, and the last applied block.