Test.Cardano.Ledger.Shelley.Examples.Init

Description

The initial state for Shelley Ledger Examples.

Documentation

ppEx ∷ (EraPParams era, ProtVerAtMost era 4, ProtVerAtMost era 6) ⇒ PParams era

Initial Protocol Parameters

initSt ∷ ∀ era. (EraTxOut era, ProtVerAtMost era 4, ProtVerAtMost era 6, Default (StashedAVVMAddresses era), EraGov era) ⇒ UTxO era → ChainState era

Initial Chain State

The initial state for the examples uses the function initialShelleyState with the genesis delegation genDelegs and any given starting UTxO set.

nonce0 ∷ ∀ c. Crypto c ⇒ Nonce

Initial Nonce

lastByronHeaderHash ∷ ∀ c. Crypto c ⇒ HashHeader c

The hash of the last Bryon Header

The first block of the Shelley era will point back to the last block of the Byron era. For our purposes in the examples we can bootstrap the chain by just coercing the value. When this transition actually occurs, the consensus layer will do the work of making sure that the hash gets translated across the fork.