Test.Cardano.Ledger.Shelley.Examples.PoolLifetime

Description

Example demonstrating the creation of a new stake pool, block production under Praos, rewards, and pool retirement.

Documentation

makePulser ∷ ∀ era. EraGov era ⇒ BlocksMade (EraCrypto era) → ChainState era → PulsingRewUpdate (EraCrypto era)

makePulser' ∷ ∀ era. EraGov era ⇒ ChainState era → PulsingRewUpdate (EraCrypto era)

makeCompletedPulser ∷ ∀ era. EraGov era ⇒ BlocksMade (EraCrypto era) → ChainState era → PulsingRewUpdate (EraCrypto era)

poolLifetimeExample ∷ TestTree

mkStake ∷ [(Credential 'Staking c, Coin)] → Stake c