Test.Cardano.Ledger.Shelley.MultiSigExamples

Documentation

applyTxWithScript ∷ ∀ c. Mock c ⇒ [(MultiSig (ShelleyEra c), Coin)] → [MultiSig (ShelleyEra c)] → Withdrawals c → Coin → [KeyPair 'Witness c] → Either (NonEmpty (PredicateFailure (ShelleyUTXOW (ShelleyEra c)))) (UTxOState (ShelleyEra c))

Start from genesis, consume Alice's and Bob's coins, create an output spendable by Alice if aliceKeep is greater than 0. For each pair of script and coin value in lockScripts create an output of that value locked by the script. Sign the transaction with keys in signers. Then create an transaction that uses the scripts in unlockScripts (and the output for aliceKeep in the case of it being non-zero) to spend all funds back to Alice. Return resulting UTxO state or collected errors

aliceOnly ∷ ShelleyEraScript era ⇒ NativeScript era

bobOnly ∷ ShelleyEraScript era ⇒ NativeScript era

aliceAndBob ∷ ShelleyEraScript era ⇒ NativeScript era

aliceOrBob ∷ ShelleyEraScript era ⇒ NativeScript era

aliceAndBobOrCarl ∷ ShelleyEraScript era ⇒ NativeScript era

aliceAndBobOrCarlAndDaria ∷ ShelleyEraScript era ⇒ NativeScript era

aliceAndBobOrCarlOrDaria ∷ ShelleyEraScript era ⇒ NativeScript era