Test.Cardano.Ledger.Shelley.Pretty

Documentation

testwidth ∷ Int → Gen a → (a → PDoc) → IO ()

Used to test pretty printing things with different widths for example: testwidth 120 ls ppLedgerState prints a random LedgerState with a max width of 120 columns one can use this to observe the how "pretty" a printer is at different widths

prettyTest ∷ TestTree