Test.Control.State.Transition.Examples.CommitReveal

Documentation

data CR hashAlgo (hashToDataMap ∷ Type → Type → Type) commitData

Commit-reveal transition system, where data hashes are committed and then revealed.

The first type parameter denotes the type of hash algorithm used.

The second parameter denotes the structure used to associate a commit to its hash. See CRSt

The third parameter is the data that will be associated to the Commit signal.

data CRSt hashAlgo hashToDataMap commitData

Commit-reveal transition system state.

Constructors

CRSt

Fields

hashToData ∷ !(hashToDataMap (Hash hashAlgo Data) commitData)

Part of the state used to associate data to the hash that was committed.

This is used only by the generators, so hashToDataMap will be instantiated to a Map in the generators, for testing purposes; and it will be instantiated to NoOpMap in an eventual implementation.

Here hashToData is an example of a phantom variable, which wouldn't be present in the formal specification, but it is needed in the executable spec to be able to generate traces.

committedHashes ∷ !(Set (Hash hashAlgo Data))

class MapLike m k v where

Methods

insert ∷ k → v → m k v → m k v

delete ∷ k → m k v → m k v

data NoOpMap a b

Constructors

NoOpMap

data CRSignal hashAlgo commitData

Constructors

Commit (Hash hashAlgo Data) commitData

Reveal Data

isCommit ∷ CRSignal hashAlgo commitData → Bool

newtype Data

Constructors

Data

Fields

getData ∷ (Id, Int)

newtype Id

Constructors

Id

Fields

getId ∷ Int

data CRPredicateFailure hashAlgo (hashToDataMap ∷ Type → Type → Type) commitData

Constructors

InvalidReveal Data

AlreadyComitted (Hash hashAlgo Data)

prop_qc_UniqueData ∷ Property

Check that unique data is generated. This is supposed to fail, since there's nothing in the STS that prevents two commits of the same data. The resulting minimal counterexample should be a trace of the form:

commit (hash d0) -> reveal d0 -> commit (hash d0)

where it shouldn't be possible to shrink d0 any further.

prop_qc_OnlyValidSignals ∷ Property

Check that only valid signals are generated.

This property should fail since the generators don't check that the generated commits are unique. The resulting minimal counterexample should be a trace of the form:

commit (hash d0) -> commit (hash d0)

where it shouldn't be possible to shrink d0 any further.