Test.Control.State.Transition.Examples.GlobalSum

Description

Simple example of a transition system whose states contain the sum of the integers seen in the signals, where sum is an abstract monoidal sum given in the enviroment.

Documentation

newtype Ops

Constructors

Ops

Fields

opSum ∷ [Int] → Int

data GSUM

newtype GSUMEvent

Constructors

ErrorEvent Void

data NoFailure

Constructors

NoFailure

spec ∷ Spec