Test.Control.State.Transition.Examples.Sum

Description

Simple example of a transition system whose states contain the sum of the integers seen in the signals.

Documentation

data SUM

data NoFailure

Constructors

NoFailure

prop_qc_Bounded ∷ Property

This property is intended to be used to manually inspect the counterexamples that we get.

prop_qc_Classified ∷ Property

Property that simply classifies the trace length distribution.

prop_qc_onlyValidSignalsAreGenerated ∷ Property