Cardano.Chain.Byron.API.Mempool

Documentation

data ApplyMempoolPayloadErr

Errors that arise from applying an arbitrary mempool payload

Although cardano-legder defines MempoolPayload, it does not define a corresponding error type. We could ChainValidationError, but it's too large, which is problematic because we actually sent encoded versions of these errors across the wire.

Constructors

MempoolTxErr UTxOValidationError

MempoolDlgErr Error

MempoolUpdateProposalErr Error

MempoolUpdateVoteErr Error

applyMempoolPayload ∷ MonadError ApplyMempoolPayloadErr m ⇒ ValidationMode → Config → SlotNumber → AMempoolPayload ByteString → ChainValidationState → m ChainValidationState

mempoolPayloadRecoverBytes ∷ AMempoolPayload ByteString → ByteString

The encoding of the mempool payload (without a AMempoolPayload envelope)

mempoolPayloadReencode ∷ AMempoolPayload a → ByteString

Re-encode the mempool payload (without any envelope)