Cardano.Chain.Genesis.Config

Documentation

data Config

Constructors

Config

Fields

configGenesisData ∷ !GenesisData

The data needed at genesis

configGenesisHash ∷ !GenesisHash

The hash of the canonical JSON representation of the GenesisData

configReqNetMagic ∷ !RequiresNetworkMagic

Differentiates between Testnet and Mainet/Staging

configUTxOConfiguration ∷ !UTxOConfiguration

Extra local data used in UTxO validation rules

data ConfigurationError

Constructors

ConfigurationGenesisDataError GenesisDataError
An error in constructing GenesisData

GenesisHashMismatch GenesisHash (Hash Raw)
The GenesisData canonical JSON hash is different than expected

GenesisHashDecodeError Text
An error occured while decoding the genesis hash.

configGenesisHeaderHash ∷ Config → HeaderHash

configK ∷ Config → BlockCount

configSlotSecurityParam ∷ Config → SlotCount

configChainQualityThreshold ∷ Fractional f ⇒ Config → f

configEpochSlots ∷ Config → EpochSlots

configProtocolMagic ∷ Config → ProtocolMagic

There isn't a full ProtocolMagic in Config, but the requisite ProtocolMagicId and RequiresNetworkMagic are stored separately. We use them to construct and return a ProtocolMagic.

configProtocolMagicId ∷ Config → ProtocolMagicId

configGenesisKeyHashes ∷ Config → GenesisKeyHashes

configHeavyDelegation ∷ Config → GenesisDelegation

configStartTime ∷ Config → UTCTime

configNonAvvmBalances ∷ Config → GenesisNonAvvmBalances

configProtocolParameters ∷ Config → ProtocolParameters

configAvvmDistr ∷ Config → GenesisAvvmBalances

mkConfigFromFile

Arguments

∷ (MonadError ConfigurationError m, MonadIO m)

⇒ RequiresNetworkMagic

→ FilePath

→ Hash Raw
The expected hash of the file

→ m Config

Construct a Config from an external genesis file.

The FilePath refers to a canonical JSON file. It will be hashed and checked against the expected hash, which should be known from config.