Cardano.Chain.Genesis.Delegation

Documentation

newtype GenesisDelegation

This type contains genesis state of heavyweight delegation. It wraps a map where keys are issuers and values are delegation certificates. There are some invariants:

In each pair delegate must differ from issuer, i. e. no revocations.

PSKs must be consistent with keys in the map, i. e. issuer's ID must be equal to the key in the map.

Delegates can't be issuers, i. e. transitive delegation is not supported. It's not needed in genesis, it can always be reduced.

Constructors

UnsafeGenesisDelegation

Fields

unGenesisDelegation ∷ Map KeyHash Certificate

data GenesisDelegationError

mkGenesisDelegation ∷ MonadError GenesisDelegationError m ⇒ [Certificate] → m GenesisDelegation

Safe constructor of GenesisDelegation from a list of Certificates