Cardano.Chain.MempoolPayload

Documentation

type MempoolPayload = AMempoolPayload ()

A payload which can be submitted into or between mempools via the transaction submission protocol.

data AMempoolPayload a

A payload which can be submitted into or between mempools via the transaction submission protocol.

Constructors

MempoolTx !(ATxAux a)
A transaction payload (transaction and witness).

MempoolDlg !(ACertificate a)
A delegation certificate payload.

MempoolUpdateProposal !(AProposal a)
An update proposal payload.

MempoolUpdateVote !(AVote a)
An update vote payload.