Cardano.Chain.Slotting.EpochNumber

Documentation

newtype EpochNumber

Index of epoch.

Constructors

EpochNumber

Fields

getEpochNumber ∷ Word64

isBootstrapEra

Arguments

∷ EpochNumber
Unlock stake epoch

→ EpochNumber
Epoch in question (for which we determine whether it belongs to the bootstrap era)

→ Bool

Bootstrap era is ongoing until stakes are unlocked. The reward era starts from the epoch specified as the epoch that unlocks stakes:

                      [unlock stake epoch]
                              /
  Epoch: ...  E-3  E-2  E-1   E+0  E+1  E+2  E+3  ...
         ------------------ | -----------------------
              Bootstrap era   Reward era