Cardano.Chain.Update.ProtocolParameters

Documentation

data ProtocolParameters

Data which is associated with BlockVersion

Constructors

ProtocolParameters

Fields

ppScriptVersion ∷ !Word16

ppSlotDuration ∷ !Natural

Milliseconds.

ppMaxBlockSize ∷ !Natural

ppMaxHeaderSize ∷ !Natural

ppMaxTxSize ∷ !Natural

ppMaxProposalSize ∷ !Natural

ppMpcThd ∷ !LovelacePortion

ppHeavyDelThd ∷ !LovelacePortion

ppUpdateVoteThd ∷ !LovelacePortion

ppUpdateProposalThd ∷ !LovelacePortion

ppUpdateProposalTTL ∷ !SlotNumber

Time to live for a protocol update proposal. This used to be the number of slots after which the system made a decision regarding an update proposal confirmation, when a majority of votes was not reached in the given number of slots. If there were more positive than negative votes the proposal became confirmed, otherwise it was rejected. Since in the Byron-Shelley bridge we do not have negative votes, and we aim at simplifying the update mechanism, ppUpdateProposalTTL is re-interpreted as the number of slots a proposal has to gather a majority of votes. If a majority of votes has not been reached before this period, then the proposal is rejected.

- TODO: it seems this should be a slot count.

ppSoftforkRule ∷ !SoftforkRule

ppTxFeePolicy ∷ !TxFeePolicy

ppUnlockStakeEpoch ∷ !EpochNumber

upAdptThd ∷ Word8 → ProtocolParameters → Int

In Byron we do not have a upAdptThd protocol parameter, so we have to use the existing ones.

lovelacePortionToRational . srMinThd . ppSoftforkRule will give us the ratio (in the interval [0, 1]) of the total stake that has to endorse a protocol version to become adopted. In genesis configuration, this ratio will evaluate to 0.6, so if we have 7 genesis keys, upAdptThd = 4.

isBootstrapEraPP ∷ ProtocolParameters → EpochNumber → Bool

Version of isBootstrapEra which takes ProtocolParameters instead of unlock stake epoch