Cardano.Chain.Update.SystemTag

Documentation

newtype SystemTag

Tag of system for which update data is purposed, e.g. win64, mac32

Constructors

SystemTag

Fields

getSystemTag ∷ Text

data SystemTagError

Constructors

SystemTagNotAscii Text

SystemTagTooLong Text

checkSystemTag ∷ MonadError SystemTagError m ⇒ SystemTag → m ()

systemTagMaxLength ∷ Integral i ⇒ i

osHelper ∷ OS → Text

Helper to turn an OS into a Text compatible with the systemTag previously used in 'configuration.yaml'

archHelper ∷ Arch → Text

Helper to turn an Arch into a Text compatible with the systemTag previously used in 'configuration.yaml'