Cardano.Chain.Update.Validation.Registration

Description

Validation rules for registering updates

This is an implementation of the rules defined in the Byron ledger specification

Documentation

data Error

Error captures the ways in which registration could fail

Constructors

DuplicateProtocolVersion ProtocolVersion

DuplicateSoftwareVersion SoftwareVersion

InvalidProposer KeyHash

InvalidProtocolVersion ProtocolVersion Adopted

InvalidScriptVersion Word16 Word16

InvalidSignature

InvalidSoftwareVersion ApplicationVersions SoftwareVersion

MaxBlockSizeTooLarge (TooLarge Natural)

MaxTxSizeTooLarge (TooLarge Natural)

ProposalAttributesUnknown

ProposalTooLarge (TooLarge Natural)

SoftwareVersionError SoftwareVersionError

SystemTagError SystemTagError

data Environment

Constructors

Environment

Fields

protocolMagic ∷ !(Annotated ProtocolMagicId ByteString)

adoptedProtocolVersion ∷ !ProtocolVersion

adoptedProtocolParameters ∷ !ProtocolParameters

appVersions ∷ !ApplicationVersions

delegationMap ∷ !Map

data State

State keeps track of registered protocol and software update proposals

Constructors

State

Fields

rsProtocolUpdateProposals ∷ !ProtocolUpdateProposals

rsSoftwareUpdateProposals ∷ !SoftwareUpdateProposals

type ApplicationVersions = Map ApplicationName (NumSoftwareVersion, SlotNumber, Metadata)

type Metadata = Map SystemTag InstallerHash

type ProtocolUpdateProposals = Map UpId (ProtocolVersion, ProtocolParameters)

type SoftwareUpdateProposals = Map UpId (SoftwareVersion, Metadata)

registerProposal ∷ MonadError Error m ⇒ Environment → State → AProposal ByteString → m State

Register an update proposal after verifying its signature and validating its contents. This corresponds to the UPREG rules in the spec.

data TooLarge n

Constructors

TooLarge

Fields

tlActual ∷ n

tlMaxBound ∷ n

newtype Adopted

Constructors

Adopted ProtocolVersion