Cardano.Chain.Update.Validation.Voting

Description

Validation rules for registering votes and confirming proposals

This is an implementation of the rules defined in the Byron ledger specification

Documentation

data Environment

Environment used to register votes and confirm proposals

Constructors

Environment

Fields

veCurrentSlot ∷ SlotNumber

veConfirmationThreshold ∷ Int

veVotingRegistrationEnvironment ∷ RegistrationEnvironment

data RegistrationEnvironment

Environment required to validate and register a vote

Constructors

RegistrationEnvironment

Fields

vreRegisteredUpdateProposal ∷ !(Set UpId)

vreDelegationMap ∷ !Map

data State

State keeps track of registered votes and confirmed proposals

Constructors

State

Fields

vsVotes ∷ !RegisteredVotes

vsConfirmedProposals ∷ !(Map UpId SlotNumber)

data Error

Error captures the ways in which vote registration could fail

Constructors

VotingInvalidSignature

VotingProposalNotRegistered UpId

VotingVoterNotDelegate KeyHash

VotingVoteAlreadyCast KeyHash

registerVoteWithConfirmation ∷ MonadError Error m ⇒ Annotated ProtocolMagicId ByteString → Environment → State → AVote ByteString → m State

Register a vote and confirm the corresponding proposal if it passes the voting threshold. This corresponds to the UPVOTE rules in the spec.