Main

Description

The main function in this file writes a file 'cardano-ledgererasalonzotest-suitesrcTestCardanoLedgerAlonzo/PlutusScripts.hs' When this file is compiled it exports a bunch of Alonzo era scripts that are plutus scripts. Compiling that file does not have any dependency on the plutus-plugin. Instead this package 'plutus-preprocessor' has that dependency, but one does not have to compile this package to build the system. If the plutus package changes, we will have to regenerate the PlutusScripts.hs file. To regenerate PlutusScripts.hs, on a machine that can depend upon plutus=plugin, then cd into the plutus-preprocessor directory and type 'cabal run plutus-preprocessor'

Documentation

display ∷ Handle → Language → ShortByteString → Q [Dec] → String → IO ()

manylines ∷ Show t ⇒ Handle → Int → [t] → IO ()

guessTheNumber'3 ∷ BuiltinData → BuiltinData → BuiltinData → ()

guessTheNumber'2 ∷ BuiltinData → BuiltinData → ()

evendata' ∷ BuiltinData → BuiltinData → BuiltinData → ()

evenRedeemer' ∷ BuiltinData → BuiltinData → BuiltinData → ()

odddata' ∷ BuiltinData → BuiltinData → BuiltinData → ()

oddRedeemer' ∷ BuiltinData → BuiltinData → BuiltinData → ()

sumsTo10' ∷ BuiltinData → BuiltinData → BuiltinData → ()

evenRedeemer2' ∷ BuiltinData → BuiltinData → ()

oddRedeemer2' ∷ BuiltinData → BuiltinData → ()

redeemerIs102' ∷ BuiltinData → BuiltinData → ()

guessTheNumberBytes ∷ ShortByteString

guess2args ∷ ShortByteString

evendataBytes ∷ ShortByteString

evenRedeemerBytes ∷ ShortByteString

odddataBytes ∷ ShortByteString

oddRedeemerBytes ∷ ShortByteString

sumsTo10Bytes ∷ ShortByteString

oddRedeemerBytes2Arg ∷ ShortByteString

evenRedeemerBytes2Args ∷ ShortByteString

redeemerIs10Bytes2Args ∷ ShortByteString

main ∷ IO ()